Separated twinses: moved `verify_commenting` to comment_util.py file
For ease of further maintainance
