+++ instapy/like_util.py
@@ -745,42 +745,3 @@ def verify_liking(browser, max, min, logger):
             return False
 
         return True
-
-
-def verify_commenting(browser, max, min, logger):
-        """ Get the amount of existing existing comments and compare it against max & min values defined by user """
-        try:
-            comments_disabled = browser.execute_script(
-                "return window._sharedData.entry_data."
-                "PostPage[0].graphql.shortcode_media.comments_disabled")
-        except WebDriverException:
-            try:
-                browser.execute_script("location.reload()")
-                comments_disabled = browser.execute_script(
-                    "return window._sharedData.entry_data."
-                    "PostPage[0].graphql.shortcode_media.comments_disabled")            
-            except:
-                logger.info("Failed to check comments' status for verification...\n")
-                raise
-                return True, 'Verification failure'
-
-        if comments_disabled == True:
-            disapproval_reason = "Not commenting ~comments are disabled for this post"
-            return False, disapproval_reason
-        try:
-            comments_count = browser.execute_script(
-                "return window._sharedData.entry_data."
-                "PostPage[0].graphql.shortcode_media.edge_media_to_comment.count")
-        except:
-            logger.info("Failed to check comments' count for verification...\n")
-            raise
-            return True, 'Verification failure'
-
-        if max is not None and comments_count > max:
-            disapproval_reason = "Not commented on this post! ~more comments exist off maximum limit at {}".format(comments_count)
-            return False, disapproval_reason
-        elif min is not None and comments_count < min:
-            disapproval_reason = "Not commented on this post! ~less comments exist off minumum limit at {}".format(comments_count)
-            return False, disapproval_reason
-
-        return True, 'Approval'
